+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * feat: Mapping to existing values in a dynamic property.
 * perf(`ObjectToObjectTransformer`): Prevent delegating to `MainTransformer` if
   the current value in a dynamic property is a scalar.
+* feat(`PresetMappingFactory`): Add `fromObjectCache()` and `fromObjectCacheReversed()`.
 
 ## 1.0.0